+++ app/Http/Controllers/AdminPhanHoiController.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
-use App\Models\KhuyenMai;
+use App\Models\AdminPhanHoi;
 use Illuminate\Http\Request;
 
-class KhuyenMaiController extends Controller
+class AdminPhanHoiController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -35,7 +34,7 @@ class KhuyenMaiController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(KhuyenMai $khuyenMai)
+    public function show(AdminPhanHoi $adminPhanHoi)
     {
         //
     }
@@ -43,7 +42,7 @@ class KhuyenMaiController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(KhuyenMai $khuyenMai)
+    public function edit(AdminPhanHoi $adminPhanHoi)
     {
         //
     }
@@ -51,7 +50,7 @@ class KhuyenMaiController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, KhuyenMai $khuyenMai)
+    public function update(Request $request, AdminPhanHoi $adminPhanHoi)
     {
         //
     }
@@ -59,7 +58,7 @@ class KhuyenMaiController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(KhuyenMai $khuyenMai)
+    public function destroy(AdminPhanHoi $adminPhanHoi)
     {
         //
     }
