+++ app/Http/Controllers/BannerController.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use App\Http\Requests\BannerRequest;
 use App\Http\Requests\SuaBannerRequest;
 use App\Models\Banner;
