+++ app/Http/Controllers/BinhLuanController.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use App\Models\BinhLuan;
 use Illuminate\Http\Request;
