+++ app/Http/Controllers/DanhGiaController.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use App\Models\DanhGia;
 use Illuminate\Http\Request;
