+++ app/Http/Controllers/KhuyenMaiSachController.php
@@ -1,21 +1,18 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
-use App\Models\DonHang;
+use App\Models\KhuyenMaiSach;
 use Illuminate\Http\Request;
 
-class DonHangController extends Controller
+class KhuyenMaiSachController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $listDonHang = DonHang::with(['sach', 'user', 'phuongThucThanhToan'])->orderByDesc('id')->get();
-
-        return view('admin.don-hang.index', compact('listDonHang'));
+        //
     }
 
     /**
@@ -37,7 +34,7 @@ class DonHangController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(DonHang $donHang)
+    public function show(KhuyenMaiSach $khuyenMaiSach)
     {
         //
     }
@@ -45,7 +42,7 @@ class DonHangController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(DonHang $donHang)
+    public function edit(KhuyenMaiSach $khuyenMaiSach)
     {
         //
     }
@@ -53,7 +50,7 @@ class DonHangController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, DonHang $donHang)
+    public function update(Request $request, KhuyenMaiSach $khuyenMaiSach)
     {
         //
     }
@@ -61,7 +58,7 @@ class DonHangController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(DonHang $donHang)
+    public function destroy(KhuyenMaiSach $khuyenMaiSach)
     {
         //
     }
