+++ app/Http/Controllers/TheLoaiController.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use App\Http\Requests\SuaTheLoaiRequest;
 use App\Http\Requests\ThemTheLoaiRequest;
 use App\Models\Sach;
