+++ app/Http/Controllers/UserController.php
@@ -1,19 +1,20 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
-use App\Models\PhuongThucThanhToan;
+use App\Models\User;
 use Illuminate\Http\Request;
 
-class PhuongThucThanhToanController extends Controller
+class UserController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        //
+        return view('admin.user.index', [
+            'users' => User::all()
+        ]);
     }
 
     /**
@@ -35,7 +36,7 @@ class PhuongThucThanhToanController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(PhuongThucThanhToan $phuongThucThanhToan)
+    public function show(string $id)
     {
         //
     }
@@ -43,7 +44,7 @@ class PhuongThucThanhToanController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(PhuongThucThanhToan $phuongThucThanhToan)
+    public function edit(string $id)
     {
         //
     }
@@ -51,7 +52,7 @@ class PhuongThucThanhToanController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, PhuongThucThanhToan $phuongThucThanhToan)
+    public function update(Request $request, string $id)
     {
         //
     }
@@ -59,7 +60,7 @@ class PhuongThucThanhToanController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(PhuongThucThanhToan $phuongThucThanhToan)
+    public function destroy(string $id)
     {
         //
     }
