+++ database/seeders/DatabaseSeeder.php
@@ -17,7 +17,7 @@ class DatabaseSeeder extends Seeder
     {
 
         for ($i = 1; $i <= 10; $i++) {
-            DB::table('user')->insert([
+            DB::table('users')->insert([
                 'ten_doc_gia' => fake()->text(20),
                 'email' => fake()->email(),
                 'mat_khau' => fake()->password(6,15),
@@ -26,6 +26,7 @@ class DatabaseSeeder extends Seeder
                 'dia_chi' => fake()->address(),
                 'sinh_nhat' => fake()->date(),
                 'gioi_tinh' => fake()->randomElement(['Nam', 'Nữ']),
+                'created_at' => now()
             ]);
         }
